requirements.txt
absl-py==1.4.0
altair==5.0.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
debugpy==1.6.7.post1
decorator==5.1.1
entrypoints==0.4
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.33
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
h5py==2.10.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
Keras-Preprocessing==1.1.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.7
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==1.1.5
parso==0.8.3
pickleshare==0.7.5
Pillow==7.0.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scikit-learn==0.22.1
scipy==1.7.3
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
